+++ middlewares/keyError.js
@@ -1,4 +1,4 @@
-const errUtils = require("../utils/errUtils");
+const errUtils = require('../utils/errUtils');
 
 const keyErrorEstate = async (req, res, next) => {
   try {
@@ -43,14 +43,14 @@ const keyErrorEstate = async (req, res, next) => {
     if (!price_main && !price_deposit) {
       throw errUtils.errGenerator({
         statusCode: 400,
-        message: "KeyError : Price",
+        message: 'KeyError : Price',
       });
     }
     // 보증금가격은 들어왔는데 월세가 없을때
     if (price_deposit && !price_monthly) {
       throw errUtils.errGenerator({
         statusCode: 400,
-        message: "KeyError : monthly",
+        message: 'KeyError : monthly',
       });
     }
 
@@ -69,4 +69,40 @@ const keyErrorEstate = async (req, res, next) => {
   }
 };
 
-module.exports = { keyErrorEstate };
+const keyErrorLogIn = async (req, res, next) => {
+  try {
+    const { email, password } = req.body;
+    if (!email || !password) {
+      throw errUtils.errGenerator({
+        statusCode: 500,
+        message: 'KeyError',
+      });
+    }
+    next();
+  } catch (err) {
+    next(err);
+  }
+};
+
+const keyErrorSignUp = async (req, res, next) => {
+  try {
+    const { email, password, username, nickname } = req.body;
+
+    const obj = { email, password, username, nickname };
+
+    for (const key in obj) {
+      if (!obj[key]) {
+        throw errUtils.errGenerator({
+          statusCode: '400',
+          message: `KEY_ERROR_${key}`,
+        });
+      }
+    }
+
+    next();
+  } catch (err) {
+    next(err);
+  }
+};
+
+module.exports = { keyErrorEstate, keyErrorLogIn, keyErrorSignUp };
