--- middlewares/validateLogin.js
@@ -1,18 +1,0 @@
-const errUtils = require('../utils/errUtils');
-
-const validateLogin = async (req, res, next) => {
-  try {
-    const { email, password } = req.body;
-    if (!email || !password) {
-      throw errUtils.errGenerator({
-        statusCode: 500,
-        message: 'KeyError',
-      });
-    }
-    next();
-  } catch (err) {
-    next(err);
-  }
-};
-
-module.exports = { validateLogin };
